requirements.txt
streamlit==1.30.0
langchain==0.1.4
langchain_core==0.1.17
langchain_community==0.0.16
langchain_google_genai==0.0.5
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
chromadb==0.3.29
fastapi==0.85.1
pydantic==1.10.21
python-dotenv==1.0.1
beautifulsoup4==4.12.2
sentence-transformers
streamlit run app.py
